Renaming to hdc-algo
move seasmon_xr into hdc.algo
