+++ README.md
@@ -1,16 +1,16 @@
-# seasmon-xr
+# hdc-algo
 Seasonal Monitor on AWS
 
-[![seasmon_xr](https://github.com/WFP-VAM/seasmon_xr/actions/workflows/seasmon_xr.yaml/badge.svg?branch=main)](https://github.com/WFP-VAM/seasmon_xr/actions/workflows/seasmon_xr.yaml)
+[![hdc-algo](https://github.com/WFP-VAM/hdc-algo/actions/workflows/main.yaml/badge.svg?branch=main)](https://github.com/WFP-VAM/hdc-algo/actions/workflows/main.yaml)
 
-For guides and How-To's, please check out the [wiki](https://github.com/WFP-VAM/seasmon_xr/wiki).
+For guides and How-To's, please check out the [wiki](https://github.com/WFP-VAM/hdc-algo/wiki).
 
 ## Installation
 
 ```bash
 # mamba/conda
-mamba install -c https://data.earthobservation.vam.wfp.org/hdc/ seasmon-xr
+mamba install -c https://data.earthobservation.vam.wfp.org/hdc/ hdc-algo
 
 # pip
-pip install --extra-index-url=https://data.earthobservation.vam.wfp.org/pypi/ seasmon-xr
+pip install --extra-index-url=https://data.earthobservation.vam.wfp.org/pypi/ hdc-algo
 ```
