+++ hdc/algo/__init__.py
