+++ hdc/algo/_version.py
