+++ hdc/algo/accessors.py
