+++ hdc/algo/ops/_helper.py
