+++ hdc/algo/ops/autocorr.py
