+++ hdc/algo/ops/lroo.py
