+++ hdc/algo/ops/spi.py
