+++ hdc/algo/ops/tinterpolate.py
