+++ hdc/algo/ops/whit.py
