+++ hdc/algo/ops/ws2d.py
