+++ hdc/algo/ops/ws2dgu.py
