+++ hdc/algo/ops/ws2doptv.py
