+++ hdc/algo/ops/ws2doptvp.py
