+++ hdc/algo/ops/ws2dpgu.py
