+++ hdc/algo/ops/ws2dwcv.py
