+++ hdc/algo/ops/ws2dwcvp.py
