+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
-name = seasmon_xr
+name =hdc-algo 
 description = xarray tools for seasonal monitor
-version = attr: seasmon_xr._version.__version__
+version = attr: hdc.algo._version.__version__
 author = WFP-VAM
 author_email =
 maintainer = WFP-VAM
@@ -10,25 +10,25 @@ long_description_content_type = text/markdown
 long_description = file: README.md
 platforms = any
 license = MIT License
-url = https://github.com/WFP-VAM/seasmon_xr/
+url = https://github.com/WFP-VAM/hdc-algo/
 project_urls =
-   Documentation = https://seasmon_xr.readthedocs.io/en/latest/
-   Bug Reporting = https://github.com/WFP-VAM/seasmon_xr/issues
+   Documentation = https://hdc-algo.readthedocs.io/en/latest/
+   Bug Reporting = https://github.com/WFP-VAM/hdc-algo/issues/
 
 classifiers =
     License :: OSI Approved :: MIT License
     Intended Audience :: Developers
     Development Status :: 2 - Pre-Alpha
     Operating System :: OS Independent
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
     Natural Language :: English
     Topic :: Software Development :: Libraries :: Python Modules
     Topic :: Scientific/Engineering :: GIS
 
 [options]
-python_requires = >=3.7
+python_requires = >=3.8
 include_package_data = false
 zip_safe = false
 packages = find_namespace:
@@ -47,7 +47,7 @@ spi =
 
 [options.packages.find]
 include =
-  seasmon_xr*
+  hdc*
 
 [aliases]
 # Define setup.py command aliases here
